# UniversityAdmission/UniversityAdmission.WebApplication/Api/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "UniversityAdmission.Server",
        "version": "1.0"
    },
    "paths": {
        "/api/Applicant": {
            "get": {
                "tags": [
                    "Applicant"
                ],
                "summary": "Получает список Абитуриентов из репозитория, в формате DtoGet",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ApplicantDtoGet"
                                    }
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ApplicantDtoGet"
                                    }
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ApplicantDtoGet"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Applicant"
                ],
                "summary": "Добавляет Абитуриента в репозиторий",
                "requestBody": {
                    "description": "добавляемая сущность в формате Dto",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ApplicantDto"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ApplicantDto"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/ApplicantDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Applicant"
                ],
                "summary": "Удаляем Абитуриента по id. А так же удаляем зависимые сущности",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/Applicant/{id}": {
            "get": {
                "tags": [
                    "Applicant"
                ],
                "summary": "Получает Абитуриента из репозитория по id, в формате Dto",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApplicantDto"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApplicantDto"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApplicantDto"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "No Content"
                    }
                }
            },
            "put": {
                "tags": [
                    "Applicant"
                ],
                "summary": "Изменяем поля Абитуриента по id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "description": "",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ApplicantDto"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ApplicantDto"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/ApplicantDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
        "/Exam": {
            "get": {
                "tags": [
                    "Exam"
                ],
                "summary": "Получает список Экзаменов из репозитория, в формате DtoGet",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ExamDtoGet"
                                    }
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ExamDtoGet"
                                    }
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ExamDtoGet"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Exam"
                ],
                "summary": "Добавляет Экзамен в репозиторий",
                "requestBody": {
                    "description": "добавляемая сущность в формате Dto",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ExamDto"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ExamDto"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/ExamDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Exam"
                ],
                "summary": "Удаляем Экзамен по id. А так же удаляем зависимые сущности",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/Exam/{id}": {
            "get": {
                "tags": [
                    "Exam"
                ],
                "summary": "Получает Экзамен из репозитория по id, в формате Dto",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExamDto"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExamDto"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExamDto"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "No Content"
                    }
                }
            },
            "put": {
                "tags": [
                    "Exam"
                ],
                "summary": "Изменяем поля Экзамена по id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "description": "",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ExamDto"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ExamDto"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/ExamDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApplicantDto": {
                "required": [
                    "city",
                    "country",
                    "dateOfBirth",
                    "fullName"
                ],
                "type": "object",
                "properties": {
                    "fullName": {
                        "type": "string",
                        "nullable": true
                    },
                    "dateOfBirth": {
                        "type": "string",
                        "format": "date"
                    },
                    "country": {
                        "type": "string",
                        "nullable": true
                    },
                    "city": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "additionalProperties": false
            },
            "ApplicantDtoGet": {
                "required": [
                    "city",
                    "country",
                    "dateOfBirth",
                    "fullName",
                    "idApplicant"
                ],
                "type": "object",
                "properties": {
                    "idApplicant": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "fullName": {
                        "type": "string",
                        "nullable": true
                    },
                    "dateOfBirth": {
                        "type": "string",
                        "format": "date"
                    },
                    "country": {
                        "type": "string",
                        "nullable": true
                    },
                    "city": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "additionalProperties": false
            },
...
            "ExamDto": {
                "required": [
                    "name"
                ],
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "additionalProperties": false
            },
            "ExamDtoGet": {
                "required": [
                    "id",
                    "name"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "name": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "additionalProperties": false
            },
...
        }
    }
}